Contractors Commands Fix (#2265)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Fix the admin command category to spawn stamps and passports
Also fix their missing ftl

---

# Media



https://github.com/user-attachments/assets/95adae52-2c10-417e-8efa-a2f57051c3b0



https://github.com/user-attachments/assets/bcf3846d-7d82-4759-8822-6e9342609ad8



# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- fix: Fixed admin commands for spawning stamps, passports, and their
related localization missing

---------

Signed-off-by: Timfa <timfalken@hotmail.com>