Fix EmptyAllContainersBehaviour Test Fail (#2277)
# Description

During a previous "Routine Cleanup", this system lost its ability to
tell the difference between a Powered Light and an AlwaysPoweredLight,
due to no longer checking if the entity has a ContainerManagerComponent.
This was essential for the function it passes the "Forced" keyword into.
The fix for this is pretty simple, just check for the component and exit
if it's not there.

Fixes this:


![image](https://github.com/user-attachments/assets/7aa9e2e3-fceb-4fdd-b55b-1cceac087ad2)


# Changelog

no CL this isn't player facing.