More Harpy Wings (#2280)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

# Description

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

About 12 new sets of harpy wings as customization options (~7 basic
options, with colorable and preset variants)

Just a few more options for harpy players to use if they want to.

---

<details><summary><h1>Media</h1></summary>
<p>


![image](https://github.com/user-attachments/assets/e372428c-6dc8-4e7b-b8d9-2873b556e6a4)

![image](https://github.com/user-attachments/assets/5e11b99e-6d58-4824-8475-dc511b0b3f0c)

![image](https://github.com/user-attachments/assets/865dcb65-3f7d-4f66-9336-98672ee015b2)

![image](https://github.com/user-attachments/assets/71e9b22e-984d-407b-bb10-d83c0c4dbee1)

![image](https://github.com/user-attachments/assets/cc6df1e8-6ac0-4613-a6f7-2e13143304a1)

![image](https://github.com/user-attachments/assets/17699b3d-22c4-42b5-a75e-a946fcd09c52)

![image](https://github.com/user-attachments/assets/d98fcf5e-bd56-43d4-b9c1-8fe666c8415e)

![image](https://github.com/user-attachments/assets/b9bfcef1-2e90-48dd-a956-771c4f2f068c)

![image](https://github.com/user-attachments/assets/08abdab8-bbc3-4bc8-8ce4-9ab16ac41a31)

![image](https://github.com/user-attachments/assets/822885b8-3abe-4e83-b394-9fee78576d3c)

![image](https://github.com/user-attachments/assets/1589c1f5-7665-4966-94a1-c84b36b2dd4f)

![image](https://github.com/user-attachments/assets/6350d409-6947-4fac-af26-84763f9ca3ef)

</p>
</details>

---

# Changelog

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl: Zeta_Null
- add: Added 12 new sets of wings for Harpies